Move internal commands to internal directory
