+++ src/commands/internal/cloneExam.js
@@ -2,10 +2,9 @@
 
 var fs            = require("fs");
 var mkdirp        = require("mkdirp");
-// var request       = require("request");
 var moment        = require("moment");
 var Promise       = require("bluebird");
-var CommandResult = require("../cli/commandResult");
+var CommandResult = require("../../cli/commandResult");
 
 var CloneChallengeCommand = require("./cloneChallenge");
